+++ src/exit.c
@@ -33,7 +33,7 @@ int bh_exit__hook_run(bh_daemon *daemon, int error_code, const char *event) {
         return_err(ret_code);
 
         if (true == run) {
-            ret_code = bh_hook_exec(daemon->paths.hooks, event, bh_err->message);
+            ret_code = bh_hook_exec(daemon->paths.hooks, event, daemon, bh_err->message);
             return_err(ret_code);
         }
 
@@ -71,7 +71,7 @@ bh_exit__hooks(bh_daemon *daemon) {
 
     /* If no other hook has been executed for the set error, we execute the
      * fatal.errors hook */
-    if (0 == runs && NULL != bh_err->message) bh_hook_exec(daemon->paths.hooks, BH_EVENT_FATALERRORS, bh_err->message);
+    if (0 == runs && NULL != bh_err->message) bh_hook_exec(daemon->paths.hooks, BH_EVENT_FATALERRORS, daemon, bh_err->message);
 }